+++ packages/pluggableWidgets/carousel-web/CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
 
+### Changed
+
+-   We updated swiper package version to 9.4
+-   DOM changes from <div> wrapper to <ul> and <li>
+
+### Fixed
+
+-   We fixed accessibility issue with carousel
+
 ## [2.1.0] - 2023-06-05
 
 ### Changed